refactor(components): [descriptions] switch to script-setup syntax
