+++ packages/components/descriptions/__tests__/descriptions.test.tsx
@@ -3,7 +3,7 @@ import { nextTick, ref } from 'vue'
 import { mount } from '@vue/test-utils'
 import { describe, expect, test } from 'vitest'
 import ElTag from '@element-plus/components/tag'
-import ElDescriptions from '../src/index.vue'
+import ElDescriptions from '../src/description.vue'
 import ElDescriptionsItem from '../src/description-item'
 
 describe('Descriptions.vue', () => {
