+++ packages/components/descriptions/index.ts
@@ -1,11 +1,14 @@
 import { withInstall, withNoopInstall } from '@element-plus/utils'
 
-import Descriptions from './src/index.vue'
+import Descriptions from './src/description.vue'
 import DescriptionsItem from './src/description-item'
 
 export const ElDescriptions = withInstall(Descriptions, {
   DescriptionsItem,
 })
-export default ElDescriptions
 
 export const ElDescriptionsItem = withNoopInstall(DescriptionsItem)
+
+export default ElDescriptions
+
+export * from './src/description'
