+++ packages/components/descriptions/src/descriptions-cell.ts
@@ -2,7 +2,7 @@
 import { defineComponent, h, inject } from 'vue'
 import { addUnit, getNormalizedProps } from '@element-plus/utils'
 import { useNamespace } from '@element-plus/hooks'
-import { elDescriptionsKey } from './token'
+import { descriptionsKey } from './token'
 
 import type { VNode } from 'vue'
 import type {
@@ -24,7 +24,7 @@ export default defineComponent({
     },
   },
   setup() {
-    const descriptions = inject(elDescriptionsKey, {} as IDescriptionsInject)
+    const descriptions = inject(descriptionsKey, {} as IDescriptionsInject)
 
     return {
       descriptions,
