+++ packages/components/descriptions/src/descriptions-row.ts
@@ -1,0 +1,8 @@
+import { buildProps } from '@element-plus/utils'
+
+export const descriptionsRowProps = buildProps({
+  row: {
+    type: Array,
+    default: () => [],
+  },
+} as const)
