+++ packages/components/descriptions/src/descriptions-row.vue
@@ -22,30 +22,21 @@
   </tr>
 </template>
 
-<script lang="ts">
+<script lang="ts" setup>
 // @ts-nocheck
-import { defineComponent, inject } from 'vue'
-import DescriptionsCell from './descriptions-cell'
-import { elDescriptionsKey } from './token'
+import { inject } from 'vue'
+
+import ElDescriptionsCell from './descriptions-cell'
+import { descriptionsKey } from './token'
+import { descriptionsRowProps } from './descriptions-row'
 
 import type { IDescriptionsInject } from './descriptions.type'
 
-export default defineComponent({
+defineOptions({
   name: 'ElDescriptionsRow',
-  components: {
-    [DescriptionsCell.name]: DescriptionsCell,
-  },
-  props: {
-    row: {
-      type: Array,
-    },
-  },
-  setup() {
-    const descriptions = inject(elDescriptionsKey, {} as IDescriptionsInject)
-
-    return {
-      descriptions,
-    }
-  },
 })
+
+defineProps(descriptionsRowProps)
+
+const descriptions = inject(descriptionsKey, {} as IDescriptionsInject)
 </script>
