+++ packages/components/descriptions/src/token.ts
@@ -1,5 +1,5 @@
 import type { InjectionKey } from 'vue'
 import type { IDescriptionsInject } from './descriptions.type'
 
-export const elDescriptionsKey =
+export const descriptionsKey =
   'elDescriptions' as any as InjectionKey<IDescriptionsInject>
